backend/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.2
django-cors-headers==4.3.1
django-environ==0.21.0
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
python-dateutil==2.8.2
requests==2.31.0
sendgrid==6.11.0
python-decouple==3.8
Pillow==10.1.0
factory-boy==3.3.0
pytest==7.4.4
pytest-django==4.7.0
pytest-asyncio==0.23.2
uuid==1.30
